chore: added s3 optimizations
